Rename asyncio module
